First push to production

+++ lib/code_corps/scenario/generate_user_mentions_for_post.rb
@@ -5,37 +5,61 @@ module CodeCorps
         @post = post
       end
 
+      attr_reader :post
+
       def call
-        result = []
-        mentions = []
-
-        @post.body.scan(/\B@((?:(?:(?:[^-\W]-?))*)(?:[^\/\W]\/?)?(?:(?:(?:[^-\W]-?))*)\w+)/) do |temp|
-          username = temp.first
-          start_index = Regexp.last_match.offset(0).first
-          end_index = start_index + username.length
-          result << [ username, [start_index, end_index] ]
-        end
-
-        users = User.where(username: result.map(&:first))
-
-        result.each do |r|
-          users.each do |u|
-            if r.first == u.username
-              mentions << [u, r.last.first, r.last.last]
-            end
-          end
-        end
-
         ActiveRecord::Base.transaction do
-          existing_mentions = PostUserMention.where(post: @post)
-          if existing_mentions.present?
-            existing_mentions.destroy_all
-          end
+          destroy_existing_mentions
           mentions.each do |m|
-            PostUserMention.create!(post: @post, user: m[0], start_index: m[1], end_index: m[2], username: m[0].username)
+            PostUserMention.create!(
+              post: @post, user: m[0],
+              start_index: m[1], end_index: m[2],
+              username: m[0].username
+            )
           end
         end
       end
-    end 
+
+      private
+
+        def destroy_existing_mentions
+          existing_mentions = post.post_user_mentions
+          existing_mentions.destroy_all if existing_mentions.present?
+        end
+
+        def regex_matches
+          regex = %r{\B@((?:(?:(?:[^-\W]-?))*)(?:[^\/\W]\/?)?(?:(?:(?:[^-\W]-?))*)\w+)}
+
+          result = []
+
+          content = post.body
+
+          return if content.nil?
+
+          content.scan(regex) do |temp|
+            username = temp.first
+            start_index = Regexp.last_match.offset(0).first
+            end_index = start_index + username.length
+            result << [username, [start_index, end_index]]
+          end
+
+          result
+        end
+
+        def mentions
+          matches = regex_matches
+          users = User.where(username: matches.map(&:first))
+
+          result = []
+
+          matches.each do |r|
+            users.each do |u|
+              result << [u, r.last.first, r.last.last] if r.first == u.username
+            end
+          end
+
+          result
+        end
+    end
   end
 end
